Add test suite through luarocks/busted
I've wanted to be able to test how the plugin interacted with neovim in
different scenarios (such as macros, `:normal`, and other ways the
plugin could indirectly be called), to ensure everything works as
expected, and that no regressions are introduced.

This adds a test suite through luarocks/busted, and a basic feature spec to
cover some of the scenarios mentioned above.